refactor: RacingManager 를 Controller 로 분리
